+++ lib/CMakeLists.txt
@@ -5,4 +5,4 @@ target_include_directories(CubeLib PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
 target_link_libraries(CubeLib PRIVATE m)
 
 find_package(Threads REQUIRED)
-target_link_libraries(CubeLib PRIVATE Threads::Threads thirdparty-cJSON)
+target_link_libraries(CubeLib PRIVATE Threads::Threads)
