+++ runner/CMakeLists.txt
@@ -2,4 +2,4 @@ cmake_minimum_required(VERSION 3.5)
 add_executable(Cube runner.c)
 target_include_directories(Cube PRIVATE ${CMAKE_CURRENT_SOURCE_DIR})
 
-target_link_libraries(Cube PRIVATE CubeLib m)
+target_link_libraries(Cube PRIVATE CubeLib m thirdparty-cJSON)
